Move Jetstream components to Components directory
The intention with this change is to signal to users that the components
are theirs for customizing.
