+++ stubs/inertia/resources/js/Components/ActionMessage.vue
