+++ stubs/inertia/resources/js/Components/ActionSection.vue
