+++ stubs/inertia/resources/js/Components/ApplicationLogo.vue
