+++ stubs/inertia/resources/js/Components/ApplicationMark.vue
