+++ stubs/inertia/resources/js/Components/AuthenticationCard.vue
