+++ stubs/inertia/resources/js/Components/AuthenticationCardLogo.vue
