+++ stubs/inertia/resources/js/Components/Banner.vue
