+++ stubs/inertia/resources/js/Components/Button.vue
