+++ stubs/inertia/resources/js/Components/Checkbox.vue
