+++ stubs/inertia/resources/js/Components/ConfirmationModal.vue
