+++ stubs/inertia/resources/js/Components/ConfirmsPassword.vue
