+++ stubs/inertia/resources/js/Components/DangerButton.vue
