+++ stubs/inertia/resources/js/Components/Dropdown.vue
