+++ stubs/inertia/resources/js/Components/FormSection.vue
