+++ stubs/inertia/resources/js/Components/Input.vue
