+++ stubs/inertia/resources/js/Components/InputError.vue
