+++ stubs/inertia/resources/js/Components/Label.vue
