+++ stubs/inertia/resources/js/Components/Modal.vue
