+++ stubs/inertia/resources/js/Components/NavLink.vue
