+++ stubs/inertia/resources/js/Components/ResponsiveNavLink.vue
