+++ stubs/inertia/resources/js/Components/SecondaryButton.vue
