+++ stubs/inertia/resources/js/Components/SectionBorder.vue
