+++ stubs/inertia/resources/js/Components/SectionTitle.vue
