+++ stubs/inertia/resources/js/Components/ValidationErrors.vue
