+++ stubs/inertia/resources/js/Layouts/AppLayout.vue
@@ -2,12 +2,12 @@
 import { ref } from 'vue';
 import { Inertia } from '@inertiajs/inertia';
 import { Head, Link } from '@inertiajs/inertia-vue3';
-import JetApplicationMark from '@/Jetstream/ApplicationMark.vue';
-import JetBanner from '@/Jetstream/Banner.vue';
-import JetDropdown from '@/Jetstream/Dropdown.vue';
-import JetDropdownLink from '@/Jetstream/DropdownLink.vue';
-import JetNavLink from '@/Jetstream/NavLink.vue';
-import JetResponsiveNavLink from '@/Jetstream/ResponsiveNavLink.vue';
+import JetApplicationMark from '@/Components/ApplicationMark.vue';
+import JetBanner from '@/Components/Banner.vue';
+import JetDropdown from '@/Components/Dropdown.vue';
+import JetDropdownLink from '@/Components/DropdownLink.vue';
+import JetNavLink from '@/Components/NavLink.vue';
+import JetResponsiveNavLink from '@/Components/ResponsiveNavLink.vue';
 
 defineProps({
     title: String,
