+++ stubs/inertia/resources/js/Pages/Dashboard.vue
@@ -1,6 +1,6 @@
 <script setup>
 import AppLayout from '@/Layouts/AppLayout.vue';
-import Welcome from '@/Jetstream/Welcome.vue';
+import Welcome from '@/Components/Welcome.vue';
 </script>
 
 <template>
